Method definition position correction
